Move test to nginx_test.go and improve test
* Improved test by allowing it to take an argument to specify which
nginx binary you want to use in the configurator
