+++ internal/nginx/configurator_test.go
@@ -851,29 +851,3 @@ func TestUpdateEndpointsMergeableIngressFailsWithInvalidTemplate(t *testing.T) {
 		t.Errorf("UpdateEndpointsMergeableIngress returned \n%v, but expected \n%v", nil, "template execution error")
 	}
 }
-
-func TestGetNginxCommand(t *testing.T) {
-	cnf, err := createTestConfigurator()
-	if err != nil {
-		t.Errorf("Failed to create a test configurator: %v", err)
-	}
-
-	cnf.nginx.nginxBinaryPath = "/usr/sbin/nginx"
-
-	tests := []struct {
-		cmd      string
-		expected string
-	}{
-		{"reload", "/usr/sbin/nginx -s reload"},
-		{"start", "/usr/sbin/nginx -s start"},
-		{"stop", "/usr/sbin/nginx -s stop"},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.cmd, func(t *testing.T) {
-			if got := cnf.nginx.getNginxCommand(tt.cmd); got != tt.expected {
-				t.Errorf("getNginxCommand returned \n%v, but expected \n%v", got, tt.expected)
-			}
-		})
-	}
-}
